Significant restructuring of pangea modules.
Reorganized pangea modules to work as alightweight wrapper
around luigi pipelines. This will result in pipelines communicating
with the server solely through the API greatly decoupling those two
codebases.

This commit is a large first step in that direction demonstrating
most of the necessary components though not yet a full pipeline.

Note that the tests here rely on a Postman Mock Server.